requirements.txt
# HINATA Phase 1 Dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.3
yfinance>=0.2.25
numpy>=1.24.3
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0